dom/webserial/tests/mochitest/test_serial_worker_main_forget.html

<!DOCTYPE HTML>
<html>
<head>
  <meta charset="utf-8">
  <title>Test Web Serial API - Main Thread Forget Propagates to Worker</title>
  <script src="/tests/SimpleTest/SimpleTest.js"></script>
  <script src="serial_test_utils.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
</head>
<body>
<script>

"use strict";

add_task(async function testMainThreadForgetPropagesToWorker() {
  await cleanupSerialPorts();

  info("Requesting port in window context");
  SpecialPowers.wrap(document).notifyUserGestureActivation();
  const port = await navigator.serial.requestPort({ filters: [] });
  ok(port, "requestPort should succeed in window context");

  let ports = await navigator.serial.getPorts();
  is(ports.length, 1, "Should have 1 port after requestPort in window");

  info("Creating worker to access granted port");
  const worker = new Worker("serial_worker_main_forget.js");

  const testPromise = new Promise((resolve, reject) => {
    worker.onmessage = async (e) => {
      if (e.data.type === "test") {
        const { result, message } = e.data;
        ok(result, message);
      } else if (e.data.type === "ready") {
        // Worker has obtained its port reference. Now forget from main thread.
        info("Worker is ready, forgetting port from main thread");
        await port.forget();

        ports = await navigator.serial.getPorts();
        is(ports.length, 0, "Main thread should have 0 ports after forget");

        // Tell the worker to check its ports.
        worker.postMessage("forgotten");
      } else if (e.data.type === "done") {
        resolve();
      } else if (e.data.type === "error") {
        reject(new Error(e.data.message));
      }
    };

    worker.onerror = (e) => {
      reject(new Error(`Worker error: ${e.message}`));
    };
  });

  try {
    await testPromise;
  } finally {
    worker.terminate();
  }
});

</script>
</body>
</html>
